Avoid test race condition.
I was seeing a race running e2e tests where immediately after applying a policy we immediately expect it to be enforced, but it takes time to compile and distribute the policy configmap.

As I was copying this other places, I noticed that the sleeps already existed, so I added my more detailed comment to the instances I found.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>