Include the predicate type and payload for attestations.
This change factors a new `PolicyAttestation` type so that we can encode additional information for attestations beyond just signature fields.

This change adds the predicate type (which would otherwise require parsing the CIP to know), and the payload (included for callers of ValidatePolicy, so they don't have to refetch the `oci.Signature`'s `Payload()`).

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>