Generate unique PolicyResult.ID for the same image.
This changes how IDs are created to use the signature digest rather than
layer digest. Because of how layers are created, this was actually a
digest of the signed content, which means that multiple signatures over
the same content would yield the same ID.

Instead, this uses a checksum of the signature, which should be unique
enough for both content + public key.

Signed-off-by: Billy Lynch <billy@chainguard.dev>